Added final features.

+++ src/pages/api/v1/users/[userNum]/delete.ts
@@ -1,4 +1,5 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
+import { decodeJWT } from "@/src/lib/authlib";
 import prisma from "@/src/lib/prisma";
 import type { NextApiRequest, NextApiResponse } from "next"
 
@@ -8,16 +9,25 @@ export default async function handler(
 ) {
 
   if (req.method == "DELETE") {
-    if (typeof req.query.userNum == "string") {
+    try {
+      const token = req.headers.authorization?.split(' ')[1];
+    
+      if (!token) {
+        return res.status(401).json({ message: 'Authentication failed: Missing token' });
+      }    
 
-      const result = await prisma.user.delete({
-        where: {
-          userNum: req.query.userNum
-        }
-      })
-        .then((resp) => console.log(`Deleted User ${ req.query.userNum }`))
+      decodeJWT(token)
 
-      res.status(200).json(result)
+      if (typeof req.query.userNum == "string") {
+        const result = await prisma.user.delete({
+          where: {
+            userNum: req.query.userNum
+          }
+        })
+      }
+      res.status(200).json({ success: "Successfully deleted user."})
+    } catch (error) { 
+      res.status(401).json({ message: 'Authentication failed: Invalid token' })
     }
   }
 }
